server/requirements.txt
python-dotenv
cassandra-driver
paho-mqtt
sympy
ariadne
pyjwt
uvicorn
quart